requirements.txt
easyocr
opencv-python
pillow
numpy
pytesseract
pdf2image
